ob-latex: htlatex and imagemagick
Hi all,

please find attached a small patch that fixes a problem discussed
in [1].

In short, when executing a latex src block and htlatex is a valid
option, it should be used even if imagemagick is part of the headers.

Regards,
Andreas

[1] http://permalink.gmane.org/gmane.emacs.orgmode/96257

>From 2986826441c24e2f0d6cf5efe6bf8aca75c4bdbb Mon Sep 17 00:00:00 2001
From: Andreas Leha <andreas.leha@gmx.de>
Date: Sat, 21 Mar 2015 22:11:37 +0000
Subject: [PATCH] ob-latex: htlatex preferred over imagemagick

* ob-latex.el (org-babel-execute:latex): change the order of the handlers

When `imagemagick' is given this entry acts as an else option.
So, it should always come last.