org.texi: Replace broken example script for batch execution
* doc/org.texi: Example bash script to tangle from *.org file was broken
and obsolete; fixed to now work with recent versions of Org.

There is an example bash script in section  14.12 of the Org manual which
is meant to tangle source blocks from a supplied list of Org files.

The present version is broken in several small ways (it appears to date
from a period when babel was still part of contrib), and requires a
hard-coded specification of the path ORG_HOME.

I've fixed the issue with ORG_HOME, another issue with grepping for strings
that no longer are emitted by ob-tangle, and removed some obsolete
dependencies.

I'm finding the corrected script a useful thing to call from makefiles.