guide: Match recommended global keybindings to manual's
* doc/org-guide.org (Activation): Match recommended global keybindings
to those in the manual.

In the guide's .texi to .org conversion in 1a678f184 (Generate compact
guide from an Org file, 2019-03-19), the leading "C-c"s of the
bindings were lost.  Restore them.

Also, drop org-switchb, which was pruned from the manual's suggested
bindings in d03515590, and tweak the surrounding text more closely
match the manual's.

Reported-by: Nick Dokos <ndokos@gmail.com>